+++ src/main/java/com/cibertec/QuickSale/model/Payment.java
@@ -2,6 +2,6 @@ package com.cibertec.QuickSale.model;
 
 public class Payment {
 
-	private long idPayment;
+	private int idPayment;
 	private String name;
 }
